added another corrections to some tasks
added another corrections to some tasks

+++ Inheritance/src/TaskNo15LoopClassInstances/CircleArrea.java
@@ -11,7 +11,7 @@ public class CircleArrea extends Formula {
 
 
 
-    public void calculate(int radius) {
+    public void calculate(double radius) {
         double pi = 3.142, area;
         area = pi * radius * radius;
         System.out.println("Area of circle is: " + area);
